fix: use decimals and not floats for currency

+++ ShoppingCart.App/Controllers/CartController.cs
@@ -32,8 +32,8 @@ namespace ShoppingCart.App.Controllers
         /// Gets the total cost value of all items within a cart.
         /// </summary>
         /// <param name="cart">An object representing a cart.</param>
-        /// <returns>A float representing a total of all cart entries within the cart.</returns>
-        public static float GetTotalCost(Cart cart)
+        /// <returns>A decimal representing a total of all cart entries within the cart.</returns>
+        public static decimal GetTotalCost(Cart cart)
         {
             return cart.CartEntries.Select(entry => CartEntryController.GetCost(entry)).Sum();
         }
